Add files via upload
This project was to predict which songs are most popular based on the given features from the dataset and to see how these factors affect what makes a song popular or unpopular on Spotify. This may provide more insight for song creators, as well the company, as to what songs or artists to feature more on their platform, ultimately generating more revenue for Spotify.